refactor(test files): modularize test files for clarity
