+++ server/config/test_config.js
@@ -1,0 +1,8 @@
+import chai from 'chai';
+import chaiHttp from 'chai-http';
+
+chai.use(chaiHttp);
+
+const { expect } = chai;
+
+export default { chai, expect };
